// examples/languages/testSoSLang/src/language-server/generated/ast.ts
export type Expr = BooleanExpression | Plus | VarRef;

export const Expr = 'Expr';

export function isExpr(item: unknown): item is Expr {
    return reflection.isInstance(item, Expr);
}

export type Statement = Assignment | Bloc | Expr | FunctionCall | If | ParallelBloc | PeriodicBloc | Variable | While;

export const Statement = 'Statement';

export function isStatement(item: unknown): item is Statement {
    return reflection.isInstance(item, Statement);
}

export interface Assignment extends AstNode {
    readonly $container: Assignment | Bloc | Conjunction | Disjunction | FunctionCall | FunctionDef | If | Model | ParallelBloc | PeriodicBloc | Plus | While;
    readonly $type: 'Assignment';
    expr: Expr
    variable: Reference<Variable>
}

export const Assignment = 'Assignment';

export function isAssignment(item: unknown): item is Assignment {
    return reflection.isInstance(item, Assignment);
}

export interface Bloc extends AstNode {
    readonly $container: Assignment | Bloc | Conjunction | Disjunction | FunctionCall | FunctionDef | If | Model | ParallelBloc | PeriodicBloc | Plus | While;
    readonly $type: 'Bloc';
    statements: Array<Statement>
}

export const Bloc = 'Bloc';
// ...
export interface FunctionCall extends AstNode {
    readonly $container: Assignment | Bloc | Conjunction | Disjunction | FunctionCall | FunctionDef | If | Model | ParallelBloc | PeriodicBloc | Plus | While;
    readonly $type: 'FunctionCall';
    args: Array<Expr>
    theFunction: Reference<FunctionDef>
}

export const FunctionCall = 'FunctionCall';

export function isFunctionCall(item: unknown): item is FunctionCall {
    return reflection.isInstance(item, FunctionCall);
}

export interface FunctionDef extends AstNode {
    readonly $container: Model;
    readonly $type: 'FunctionDef';
    body: Bloc
    name: string
    params: Array<string>
}

export const FunctionDef = 'FunctionDef';

export function isFunctionDef(item: unknown): item is FunctionDef {
    return reflection.isInstance(item, FunctionDef);
}

export interface If extends AstNode {
    readonly $container: Assignment | Bloc | Conjunction | Disjunction | FunctionCall | FunctionDef | If | Model | ParallelBloc | PeriodicBloc | Plus | While;
    readonly $type: 'If';
    cond: VarRef
    else: Bloc
    then: Bloc
}

export const If = 'If';

export function isIf(item: unknown): item is If {
    return reflection.isInstance(item, If);
}

export interface Model extends AstNode {
    readonly $type: 'Model';
    functionDefs: Array<FunctionDef>
    statements: Array<Statement>
}

export const Model = 'Model';

export function isModel(item: unknown): item is Model {
    return reflection.isInstance(item, Model);
}

export interface ParallelBloc extends AstNode {
    readonly $container: Assignment | Bloc | Conjunction | Disjunction | FunctionCall | FunctionDef | If | Model | ParallelBloc | PeriodicBloc | Plus | While;
    readonly $type: 'ParallelBloc';
    statements: Array<Statement>
}

export const ParallelBloc = 'ParallelBloc';

export function isParallelBloc(item: unknown): item is ParallelBloc {
    return reflection.isInstance(item, ParallelBloc);
}

export interface PeriodicBloc extends AstNode {
    readonly $container: Assignment | Bloc | Conjunction | Disjunction | FunctionCall | FunctionDef | If | Model | ParallelBloc | PeriodicBloc | Plus | While;
    readonly $type: 'PeriodicBloc';
    bloc: Bloc
    time: number
}

export const PeriodicBloc = 'PeriodicBloc';
// ...
export interface Variable extends AstNode {
    readonly $container: Assignment | Bloc | Conjunction | Disjunction | FunctionCall | FunctionDef | If | Model | ParallelBloc | PeriodicBloc | Plus | While;
    readonly $type: 'Variable';
    initialValue?: number
    name: string
}

export const Variable = 'Variable';
// ...
export interface While extends AstNode {
    readonly $container: Assignment | Bloc | Conjunction | Disjunction | FunctionCall | FunctionDef | If | Model | ParallelBloc | PeriodicBloc | Plus | While;
    readonly $type: 'While';
    body: Bloc
    cond: VarRef
}

export const While = 'While';